Frisbee/requirements.txt
adium-theme-ubuntu==0.3.4
ansible==2.0.0.2
attrs==15.2.0
beautifulsoup4==4.4.1
blinker==1.3
ccsm==0.9.12.2
chardet==2.3.0
command-not-found==0.3
compizconfig-python==0.9.12.2
configglue==1.1.2
cryptography==1.2.3
debtags==2.0
defer==1.0.6
dirspec==13.10
Django==1.9.2
django-grappelli==2.9.1
django-nested-admin==2.1.12
django-suit==0.2.15
django-super-inlines==0.1.4
django-tables2==1.1.1
dnspython==1.12.0
duplicity==0.7.6
ecdsa==0.13
enum34==1.1.2
f.lux-indicator-applet==1.1.8
gyp==0.1
html5lib==0.999
httplib2==0.9.1
idna==2.0
ipaddress==1.0.16
Jinja2==2.8
keyring==7.3
launchpadlib==1.10.3
lazr.restfulclient==0.13.4
lazr.uri==1.0.3
lockfile==0.12.2
lxml==3.5.0
Mako==1.0.3
MarkupSafe==0.23
meld==3.14.2
mysql-connector-python==2.0.4
mysql-utilities==1.6.1
mysqlclient==1.3.7
ndg-httpsclient==0.4.0
netaddr==0.7.18
numpy==1.11.0
oauth==1.0.1
oauthlib==1.0.3
oneconf==0.3.9
PAM==0.4.2
paramiko==1.16.0
pexpect==4.0.1
Pillow==3.1.2
piston-mini-client==0.7.5
protobuf==2.6.1
ptyprocess==0.5
pyasn1==0.1.9
pyasn1-modules==0.0.7
pycrypto==2.6.1
pycups==1.9.73
pycurl==7.43.0
pygobject==3.20.0
pyinotify==0.9.6
PyJWT==1.3.0
pyodbc==3.0.7
pyOpenSSL==0.15.1
pyserial==3.0.1
pysmbc==1.0.15.5
pysqlite==2.7.0
python-apt==1.1.0b1
python-debian==0.1.27
python-monkey-business==1.0.0
pyxdg==0.25
PyYAML==3.11
reportlab==3.3.0
requests==2.9.1
SecretStorage==2.1.3
service-identity==16.0.0
simplejson==3.8.1
six==1.10.0
software-center-aptd-plugins==0.0.0
stevedore==0.14.1
termcolor==1.1.0
Twisted==16.0.0
ubuntuone-client-data==14.4
uniconvertor==1.1.5
urllib3==1.13.1
vboxapi==1.0
virtualenv==15.0.1
virtualenv-clone==0.2.4
virtualenvwrapper==4.2
wadllib==1.3.2
youtube-dl==2016.11.18
zope.interface==4.1.3
